Extend JAXBResourceFactory tests to confirm that post construct is called on the jaxb type if it implements it
